% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Human-PfSI.R
\name{check_human_pfsi_conpars}
\alias{check_human_pfsi_conpars}
\title{Human-PfSI: Check Constructor Parameters for a Population}
\usage{
check_human_pfsi_conpars(par)
}
\arguments{
\item{par}{a list}
}
\description{
Given a list where each element is output from \code{\link{human_pfsi_conpars}},
double check to make sure all elements are valid constructor parameters for that human.
}
