% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PfSI-Parameters.R
\name{pfsi_parameters}
\alias{pfsi_parameters}
\title{PfSI Parameters}
\usage{
pfsi_parameters(Pf_c = 0.15, Pf_b = 0.55, DurationPf = 200,
  LatentPf = 10, FeverPf = 0.3, mnFeverPf = 10, vrFeverPf = 0.1,
  TreatPf = 0.5, mnTreatPf = 1, mnChemoprophylaxisPf = 32,
  PEProtectPf = 0.99, peBlockPf = 1, mnPEPf = 270, vrPEPf = 50,
  GSProtectPf = 1, gsBlockPf = 0.9, mnGSPf = 180, vrGSPf = 20)
}
\arguments{
\item{Pf_c}{human to mosquito transmission efficiency}

\item{Pf_b}{mosquito to human transmission efficiency}

\item{DurationPf}{duration of infection}

\item{LatentPf}{duration of latent stage}

\item{FeverPf}{probability of fever upon infection}

\item{mnFeverPf}{mean of time to fever (Gaussian)}

\item{vrFeverPf}{sd of time to fever (Gaussian)}

\item{TreatPf}{probability of seeking treatment after fever}

\item{mnTreatPf}{time to treatment}

\item{mnChemoprophylaxisPf}{duration of chemoprophylactic protection}

\item{PEProtectPf}{vaccination efficacy (probability of effect)}

\item{peBlockPf}{vaccine effectiveness (reduction in probability of transmission 'b' given challenge)}

\item{mnPEPf}{mean duration of protection from PE vaccination (Gaussian)}

\item{vrPEPf}{standard deviation of protection from PE vaccination (Gaussian)}

\item{GSProtectPf}{vaccination efficacy (probability of effect)}

\item{gsBlockPf}{vaccine effectiveness (reduction in probability of transmission 'c' given challenge)}

\item{mnGSPf}{mean duration of protection from GS vaccination (Gaussian)}

\item{vrGSPf}{standard deviation of protection from GS vaccination (Gaussian)}
}
\description{
Parameters for the 'PfSI' model of human infection dynamics
(see inst/include/Parameters.hpp for where parameters are stored)
}
